% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_humidity_month.R
\name{history_humidity_month}
\alias{history_humidity_month}
\title{Plot_humidity_monthly}
\usage{
history_humidity_month(df = history_weather_tbl)
}
\description{
plot historical humidity by month
             1) Mean humidity
             2) Max humidity
             3) Min humidity
}
\examples{
history_weather_tbl=history_weather(city_name = "Alameda",start_date="2018-06-01",end_date="2018-12-31")
history_humidity_month(data=history_weather_tbl)
}
\keyword{by}
\keyword{humidity}
\keyword{month}
