% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_daily.R
\name{history_temp_daily}
\alias{history_temp_daily}
\title{Plot_temperature_daily}
\usage{
history_temp_daily(df = history_weather_tbl)
}
\description{
plot historical temperature by date
             1) Mean temperature
             2) Max temperature
             3) Min temperature
}
\examples{
history_weather_tbl=history_weather(city_name = "Alameda",start_date="2018-06-01",end_date="2018-12-31")
history_temp_daily(data=history_weather_tbl)
}
\keyword{by}
\keyword{date}
\keyword{temperature}
