% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_month.R
\name{history_temp_month}
\alias{history_temp_month}
\title{Plot_temperature_monthly}
\usage{
history_temp_month(df = history_weather_tbl)
}
\description{
plot historical temperature by month
             1) Mean temperature
             2) Max temperature
             3) Min temperature
}
\examples{
history_weather_tbl=history_weather(city_name = "Alameda",start_date="2018-06-01",end_date="2018-12-31")
history_temp_month(data=history_weather_tbl)
}
\keyword{by}
\keyword{month}
\keyword{temperature}
