% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_weather.R
\name{history_weather}
\alias{history_weather}
\title{history weather}
\usage{
history_weather(city_name = "Belmont", start_date = "2018-05-05",
  end_date = "2018-06-05")
}
\description{
given city and date range, create table of historical weather for local PWS.
}
\examples{
history_weather_tbl=history_weather(city_name = "Alameda",start_date="2018-06-01",end_date="2018-12-31")
}
\keyword{history_weather}
