% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distm_city.R
\name{nearby_pws_city}
\alias{nearby_pws_city}
\title{List all PWS nearby the city selected}
\usage{
nearby_pws_city(city_name = "Belmont", check_date = "2018-05-05")
}
\description{
given a city name in Bay area, select all PWS with weather data in history
}
\examples{
pws_id_selected <- nearby_pws_city(city_name = "Alameda",check_date="2018-05-05")
}
\keyword{nearby_pws_city}
