% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distm_coordinates.R
\name{nearby_pws_coordinates}
\alias{nearby_pws_coordinates}
\title{List all PWS nearby given coordinates and distance}
\usage{
nearby_pws_coordinates(lon = -122.2, lat = 37.5, distance = 2000,
  data = pws, check_date = "2018-05-05")
}
\description{
given coordinates and distance in Bay area, select all PWS with weather data in history
}
\examples{
pws_id_selected=nearby_pws_coordinates(lon=-122.27999878,lat=37.52000046,distance = 4000,data=pws,check_date = "2018-05-05")
}
\keyword{nearby_pws_coordinates}
