% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pws.R
\docType{data}
\name{pws}
\alias{pws}
\title{pws data for 290 final project}
\format{a large list (3560 elements, 10.3Mb):
\describe{
  \item{}{}
  \item{}{}
  ...
}}
\source{
\url{http://www.stanford.edu/}
}
\usage{
pws
}
\description{
pws data for 290 final project
}
\keyword{datasets}
