% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWeatherUI.R
\name{runWeatherUI}
\alias{runWeatherUI}
\title{run shiny WeatherUI}
\usage{
runWeatherUI()
}
\description{
run shiny WeatherUI
}
\examples{
\dontrun{
library(ggmap)
library(gridExtra)
stat290.finalproject::runWeatherUI()
}


}
