% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{weather_map}
\alias{weather_map}
\title{Create a ggplot map using google map}
\usage{
weather_map(df = pws_id_selected)
}
\description{
1) Generate a map based on the selected list of personal weather stations(pws)
             2) plot contour lines based on tempatures in the neighborhood
}
\examples{
source("distm_city.R")
pws_id_selected <- nearby_pws_city(city_name = "Alameda",check_date="2018-05-05")
weather_map(df=pws_id_selected)
}
\keyword{weather_map}
